refactor: move validateRateLimit to reusable module
